iommu/exynos: add support for IOMMU mapping for virtual devices
IOMMU mapping can be shared across devices so that multiple devices
can access the same page table base. This sharing can be across
devices which are virtual (logical) like DRM which share their mapping
with DRM-FIMD/DRM-HDMI etc.
This patch enables IOMMU mapping creation for a virtual device which
does not have an actual SYSMMU. The iommu driver checks if a sysmmu
exists and returns if it is not present.

Change-Id: I39b8d3ee6e6a10a84c37570b0f67dbad7c154a05
Signed-off-by: Prathyush K <prathyush.k@samsung.com>