UPSTREAM: DMA: PL330: Fix racy mutex unlock
pl330_update() stores a pointer to the thrd->req that finished, which
contains a pointer to the corresponding pl330_req.  This is done with
the pl330_lock held.  Then, it iterates through the req_done list,
calling the callback for each of the requests that are done.  The
problem is that the driver releases the lock before calling the
callback for each of the callbacks.  pl330_submit_req() running in
another processor can then acquire the lock and insert another request
in one of the thrd->req that hasn't been processed yet, replacing the
pointer to pl330_req there.  When the callback returns in
pl330_update() and the next rqdone is popped from the list, it
dereferences the pl330_req pointer to the just scheduled pl330_req,
instead of the one that has finished, calling pl330 with the wrong r.

This patch fixes this by storing the pointer to pl330_req directly in
the list.

BUG=chrome-os-partner:12565
TEST=built for emerge-daisy chromeos-kernel and booted on one board.

Signed-off-by: Javi Merino <javi.merino@arm.com>
Cc: Jassi Brar <jaswinder.singh@linaro.org>
Acked-by: Jassi Brar <jaswinder.singh@linaro.org>
Signed-off-by: Vinod Koul <vinod.koul@linux.intel.com>
Signed-off-by: Alim Akhtar <alim.akhtar@gmail.com>
(cherry picked from commit fdec53d5203ee3d44be7c09d1524e3a6287d46a7)

Change-Id: If0f7f4dd93dc12acf38047c8d90406bd172eb875
Reviewed-on: https://gerrit.chromium.org/gerrit/31829
Commit-Ready: naveen krishna chatradhi <naveen@chromium.org>
Tested-by: naveen krishna chatradhi <naveen@chromium.org>
Reviewed-by: Hatim Ali R.V <hatim.rv@samsung.com>
Reviewed-by: naveen krishna chatradhi <naveen@chromium.org>