rpmsg: resmgr: omap: retrieve ISS opt clk handle from platform data
The OMAP rpmsg driver is currently getting the struct clk handle for
the ISS optional clock through the omap_get_clk_by_name function. This
function is currently not exported, and causes a build error if
omap_rpmsg_resmgr were to be built as a module. This function call is
moved to the mach-omap2/rpmsg_resmgr.c which is always built-in, and
the ISS optional clock struct clk is published to the generic rpmsg
omap resmgr layer through platform data to its omap driver.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>