FIX: add blank mmap function support for dma_buf_ops
In the new dma_buf_export(), there is a specific check for support
ops->mmap. vb2 patches are not supplying this at the time of writing
this commit. Hence a empty function is added to support. AFAIK, there
is no need to supply mmap support for vb2, as REQBUF already does it
if buffers are allocated by driver. But when the buffers are imported
there is no support, and not sure of any use case as of now.

Change-Id: I2f457ff1b24135eeb2910676eb2693143747f582
Signed-off-by: Subash Patel <subash.rp@samsung.com>