drm/exynos: Add exynos drm specific fb_mmap function
This fb_mmap function is called from the fb_mmap default function
in fbmem.c. This new function does not assume that the frame buffer
is contiguous and calls dma_mmap_writecombine for mapping the buffer
to user space. dma_mmap_writecombine will be able to map a contiguous
buffer as well as non-contig buffer.

Change-Id: I380927c1ad5c36069135b5868bfe9af8e0d1a54e
Signed-off-by: Prathyush K <prathyush.k@samsung.com>