irqchip: Add ARM Generic Interrupt Controller support
This patch adds support for ARM GICv2 under drivers/irqchip/. It is
expected that ARMv8 systems will use GICv3 and this code may be
deprecated in the future. The GIC initialisation is driven by the FDT.
And the secondary CPU interface is initialised via a CPU notifier to
allow late SoC initialisation.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>