v4l: vb2: unmap dmabufs on STREAM_OFF event
[ -- patch under review upstream -- ]

Currently, if the user issues a STREAM_OFF request and then
tries to re-enqueue buffers, it will trigger a warning in
the vb2 allocators as the buffer would still be mapped
from before STREAM_OFF was called. The current expectation
is that buffers will be unmapped in dqbuf, but that will never
be called on the mapped buffers after a STREAM_OFF event.

Change-Id: I76c9db160688819490f51884bf7fc8e2dd25229b
Signed-off-by: Dima Zavin <dima@android.com>