ARM: exynos: dt: turn on wwan-en
This turns on wwan all the time which is fine at the moment.
When we want to optimize for power, we'll need to figure out how
to hook things in such that:
* We turn off 3G when the user has disabled it in the UI (that
  would mean we couldn't tell if we're on a 3G sku).
* We turn off 3G during suspend (assuming that we don't want to
  keep it on for some reason).
See the bug for more details.

BUG=chrome-os-partner:10528
TEST=Before this change, 'grep "wwan-en" /sys/class/regulator/*/name'
returned nothing.  After this change, it showed a match.
TEST=Before this change, saw no 3G module in 'lsusb'.  After this
change I saw a 3G module in lsusb.

Change-Id: I34b12cf0ad5d30d246c31aa707f32a16c5351ca8
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/20508
Tested-by: Doug Anderson <dianders@google.com>
Reviewed-by: Olof Johansson <olofj@google.com>
Commit-Ready: Doug Anderson <dianders@google.com>
Reviewed-by: Terry Lambert <tlambert@google.com>