HID: magicmouse: Correct report range of major / minor axes
In patch "HID: magicmouse: Adjust major / minor axes to scale",
touch_major and touch_minor axes are scaled by a factor of
four when reported but the max touch_major/minor is not scaled
accordingly. This patch scales the max touch_major/minor to
be consistent with the reported value.

BUG=chromium-os:29435
TEST=Use evtest to test Apple Magic Trackpad and check that the
     range for touch_major/minor is [0 1020] and the reported
     touch_major/minor value is within the range.

Change-Id: Id1d54e1881025c734e908f90ddcb8e0f12c29d88
Signed-off-by: Yufeng Shen <miletus@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25844
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>