CHROMIUM: ARM: exynos: handle setaffinity on combined interrupts
This is needed to set affinity of the PMU interrupts for perf. Note that
the whole group of combined interrupts will be moved to whatever the last
affinity was set to. This might or might not cause problems, but it's no worse
than now.

BUG=none
TEST=run perf under taskset -c 0/1 and see it still work

Change-Id: I76b0aa00e040c5f0490be3e4c885dfd2d20e1d79
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/20680
Reviewed-by: Sonny Rao <sonnyrao@google.com>
Commit-Ready: Olof Johansson <olofj@google.com>
Tested-by: Olof Johansson <olofj@google.com>