gpio: samsung: add flags specifier to device-tree binding
This adds a flags field to the gpio specifier for Samsung. I didn't
want to add yet another field in the already quite long specifier, so
I decided to compress it together with the Pull Up/Down settings instead.

This is needed to, for example, have a gpio-keys input that is active low.

BUG=chrome-os-partner:10861
TEST=boot, make sure system stays awake, close lid, make sure system suspends

Change-Id: Ia3ad27c13ab773da6d584370ac585d7c982590a5
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/20819
Reviewed-by: Doug Anderson <dianders@google.com>
Reviewed-by: <thomas.ab@samsung.com>
Tested-by: Olof Johansson <olofj@google.com>
Commit-Ready: Olof Johansson <olofj@google.com>