drm/exynos: wait for vsync before removing FB
While removing a framebuffer, we wait for a vsync interrupt. This is
required if a new framebuffer has been just set and has not yet been
updated to the actual registers. In this case, if we remove the older
buffer, we get a page fault, since the older buffer is still being used.
A framebuffer can be safely removed only after receiving a vsync.

This is also required when we try to remove the current framebuffer.
FIMD gets disabled but has to wait for a vsync interrupt before the
disabling takes effect.

BUG=chrome-os-partner:10728, chrome-os-partner:10813
TEST='stop ui/pkill -9 X' and no page fault. Tested on Daisy/Snow.

Change-Id: Ib779526b7d116a8f546178e4fbd1a7d036be2771
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/26382
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Commit-Ready: Doug Anderson <dianders@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>