video: s3c-fb: set default alpha to (255, 255, 255)
Previously overlays defaulted to ALPHA0_{R,G,B} = (0, 0, 0) which was
used when the buffer had no alpha channel.  An alternate
ALPHA1_{R,G,B} = (240, 240, 240) was configured but not used.  This
caused RGB-565 overlays to effectively draw in reverse Z-order.

Alpha-related register definitions are also corrected and/or simplified
to handle the various methods of configuring ALPHA{0,1}.

Signed-off-by: Greg Hackmann <ghackmann@google.com>

Change-Id: I2f0363f3a29a28fc635009883c5e885c87756580