HACK: ARM: Add option to disable branch prediction during decompress
This is a HACK because we are always enabling the Z bit after
decompression instead of restoring to the original value.

The early (pre-decompress) mmu (page table) setup for v7 is minimal
and does not have any knowledge of device memory or secure regions
of memory. Cortex-A15 has a beefier instruction prefetch among other
things (deeper btb..) and is causing accesses to secure regions at
this point in time.

This new config option adds ability to disable the Z bit and get rid
of these bad accesses. This Z bit is then enabled after decompression.

Change-Id: I59c77f979b174957a74ceb3f5ff93fe2050ff507
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>
Signed-off-by: Marquette Anderson <m-anderson13@ti.com>
Signed-off-by: Vincent Stehlé <v-stehle@ti.com>