CHROMIUM: exynos: dvfs: maxim: Fix unit conversion error when calculating delay
The Maxim 77686 driver specifies voltages for different regulators in different
units, using either milli-volts or micro-volts.  It fails to account for this
unit difference when calculating the amount of time to wait in microseconds for
the voltage change to complete, so I changed the table of slew rates to be in
micro-volts per micro-second instead of milli-volts per micro-second, and
factor in the units into the calculation.  I also made the table more
accurate than it was because we have more precision by using micro-volts.

BUG=chrome-os-partner:10891
TEST="ad hoc, run with ondemand or interactive governer, run perf top,
observe that __delay is no longer one of the top functions"

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: Ic2ae82d92a86db503982b6cc29fdc2d68fa6ee9e
Reviewed-on: https://gerrit.chromium.org/gerrit/26851
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>