CHROMIUM: exynos: dvfs: fix default slew rate for Maxim 77686 PMIC
The device tree specifies the default voltage slew rate for the
Maxim 77686 regulator which is powering the CPU rail, but it is giving
the incorrect value.  The value isn't well specified as it merely points
to a table in the regulator code, but the current value of 2 points to the
55 mv/usec rate instead of the actual default which is 27.5 mv/usec.
This happened to work because the maxim regulator code was incorrectly
ignoring the value coming from the device tree, and choosing the
appropriate default.  Fix the first part by changing the default value
to 1 which is 27.5 mV/uS.

BUG=chrome-os-partner:10891
TEST=none, requires further driver fix to test

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: I192a24d96f0642c58f555c89061df0a5e86508d9
Reviewed-on: https://gerrit.chromium.org/gerrit/26852
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>