arm: omap: devices: make softreset when deactivating iss and fdif
ISS and FDIF are specialized accelerators and are used by the IPU
subsystem. These need to be a proper state both during regular usage
as well as across remoteproc recoveries. These should be put under
reset ideally when they are deactivated to ensure a  successful
shutdown and a proper state for further usages. However, ISS and
FDIF  are part of the CAM domain and do not have an associated hard
reset line. So, a softreset is performed just before deactivating ISS
and FDIF modules.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>