CHROMIUM: drm/i915: Adjust the RPS thresholds
Lower the RPS thresholds and make them symmetric. This patch increases
the rate of change of the GPU clock, allowing it to clock up more
quickly, and more importantly, clock down when necessary. The problem
this patch solves is that a heavy workload would inflate the clock, and
then the clock would stick high when the workload went away because the
down threshold was too high. This should improve our thermals and power.

BUG=chrome-os-partner:11009
TEST=Tested with the PBS koala video, play it along with a GL demo to
get the clock up to 1150MHz. Remove the GL demo workload and ensure the
clock goes back down to 350MHz.

Change-Id: I9a489ffd4810fb2b3c2544d6165d664899ec0e54
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27184
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>