omap: mailbox: fix api used by mailbox to hold a latency constraint
dev_pm_qos API are currently used to hold a latency constraint on
mailbox, which is present in the core power domain. This effectively
holds a constraint on the core power domain.

However, the pm core uses pm_qos APIs to control the latency of the
core domain and thus the C-states in OMAP. The two sets of interfaces
(pm_qos and dev_pm_qos) are de-coupled and thus cannot be used at the
same time to control latency values for core domain.

The mailbox latency code has therefore been modified to use the
pm_qos API for holding latency constraints, so that the pm core
code has a consistent view of the constraints set on core power domain.

Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>