drm/exynos: HACK: return an error when flip called with event flag
The event flag to the page_flip ioctl is optional. If it is set,
an event will be sent to the drm fd after page flip is complete.
Otherwise, no event is sent after page flip. The exynos implementation
was doing nothing if the event flag was not set. This would
break any application expecting a page flip to happen. Rather than
do nothing (which is incorrect behavior), we'll now return and log
an error.

Rather than silently breaking applications, we'll return and log
an error so that we can detect anything that's broken.

This also removes one level of indenting, making the code easier to
read.

BUG=none
TEST=Verified that no errors were logged.

Change-Id: I31af40a392c89ea4386bfe38c00d79979aa9dc0b
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28853