ARM: exynos: dp: Move bridge init into phy init
Move the eDP bridge initialization code into the dp phy init function so
we run it both on startup and on resume. As a side-effect of this
change, we need to wait 90ms so we know the bridge chip has come out of
reset since the phy init is much closer to the HPD loop.

BUG=chrome-os-partner:11908
TEST=Tested on DVT and PVT snow

Change-Id: I3522b8aacd4a8cebc8bc230e81fe23b9afae02f2
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28626
Reviewed-by: Doug Anderson <dianders@chromium.org>