remoteproc: add a new 'recover' option to recovery debugfs
The remoteproc infrastructure allows to disable recovery using a
debugfs entry by doing the following:
	echo disabled > /d/remoteproc/remoteprocX/recovery

This is desired during debugging to inspect the state of memory
upon a remote processor crash. However, there is no provision
currently to trigger recovery after disabling it in debugfs.

A new option "recover" can be written to 'recovery' debugfs entry
of the remoteproc to perform a one-time recovery explicitly in
these conditions. This option is a no-op if recovery is not
disabled or if remoteproc has not already crashed.

The usage is as follows:
	echo recover > /d/remoteproc/remoteprocX/recovery

Change-Id: I7fc130e40c64ab22f1893b493fd72086393d86a4
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>