usb: musb: fix suspend/resume after gadget driver is loaded and unloaded
If the MUSB controller is already suspended via runtime suspend then it
will be disabled (and clocks gated) thus accessing the controller
while system suspend to save state will cause L3 bus errors.

This problem can be reproduced by loading a gadget driver, unloading
it and putting the system to suspend.

There is no need to save/restore controller state in this case
as the corresponding runtime suspend/resume functions are already doing that.

This patch makes sure we don't access the MUSB controller registers in the
suspend path if it is already disabled.

Signed-off-by: Roger Quadros <rogerq@ti.com>