CHROMIUM: regulator: Add device tree node for max77686
This just passes the device tree node to regulator_register() which
allows get_regulator() to work when supplies are defined in the device
tree.

NOTE: The maxim77686 driver has changed significantly upstream so this
is not easily posted upstream.  Hopefully the code there has already
been fixed.

BUG=chrome-os-partner:11688
TEST=In follow-on CLs I have specified supplies in the device tree.
Before this change that didn't work for max77686 provided supplies.
After this change that works.

Change-Id: I93a46db7e80910a8a98ac89c93e1291b96f97dad
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28983
Reviewed-by: Olof Johansson <olofj@chromium.org>