CHROMIUM: cfg80211: Clear "beacon_found" on regulatory restore
Restore the default state to the "beacon_found" flag when
the channel flags are restored.  Otherwise, we can end up
with a channel that we can no longer transmit on even when
we can see beacons on that channel.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:12087
TEST=Rerun network_WiFiMatFunc.004CheckHiddenSSID 10 times.
Monitor channel flags in "iw list" for a passive-only channel.

Change-Id: Ief2762e08d236a0ac60cb00b96eafc3b23af75f2
Reviewed-on: https://gerrit.chromium.org/gerrit/28996
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>