CHROMIUM: x86: smp: don't send NMI while stopping CPUs
There have been an increasing number of reports of systems freezing.
The kernel watchdogs are not causing reboots.

While debugging NMI watchdog issues couple of months back I had noticed
that the new code that sends NMIs to other CPUs can cause lockups since
the perf counter NMI handler also gets executed in this path.  The perf
counter handler is a _lot_ of code to be executing when the CPUs are
being stopped especially in panic situations.

Setting smp_no_nmi_ipi to true to see if this helps bring down the
system freezing reports and also hopefully provides us stack traces to
debug the root cause of the panics.

BUG=chromium-os:32987
TEST="echo bug > /proc/breakme" and check that kcrash files are
generated correctly.

Change-Id: I276ea6c6918ab05714692bd614e7dfe179a4d32a
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29085
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>